src/loader/app/requirements.txt
azure-storage-blob
decouple